requirements.txt
SpeechRecognition
pyttsx3==2.90
wikipedia
PyAudio
sounddevice==0.5.2
numpy==2.3.4
piper-tts